ir: Decouple IterationProperty and Tag
