+++ devito/tools.py
@@ -1,3 +1,4 @@
+from abc import ABC
 import os
 import inspect
 import ctypes
@@ -702,6 +703,44 @@ class ReducerMap(MultiDict):
         Returns a dictionary with reduced/unique values for all keys.
         """
         return {k: self.reduce(key=k) for k in self}
+
+
+class Tag(ABC):
+
+    """
+    An abstract class to define categories of object decorators.
+
+    .. note::
+
+        This class must be subclassed for each new category.
+    """
+
+    _repr = 'AbstractTag'
+
+    _KNOWN = []
+
+    def __init__(self, name, val=None):
+        self.name = name
+        self.val = val
+
+        self._KNOWN.append(self)
+
+    def __eq__(self, other):
+        if not isinstance(other, self.__class__):
+            return False
+        return self.name == other.name and self.val == other.val
+
+    def __hash__(self):
+        return hash((self.name, self.val))
+
+    def __str__(self):
+        return self.name if self.val is None else '%s%s' % (self.name, str(self.val))
+
+    def __repr__(self):
+        if self.val is None:
+            return "%s: %s" % (self._repr, self.name)
+        else:
+            return "%s: %s[%s]" % (self._repr, self.name, str(self.val))
 
 
 # Method/function arguments validation
